Screenrecord: Allow to select AVC codec profile level
Nevertheless we have an option to limit the recording by frame rate, but
we still can't limit it by resolution. It's not particularly needed for
the devices of today which are enough powerful and can encode/play any
profile level, but there are still many old devices for which the
profile level can make a big difference.

Traditionally the device declares the high quality multimedia profiles
for video encoding (editing, recording video via camera ecc.), but these
profiles are also used for screen recording. And here comes the pitfall
for weak devices: screen recording is a video encoding process +
highload user tasks on foreground and background, all in total is too
much for an old device. The recorded video is "laggy", because the
frames are simply skipped to the detriment of system smoothness. That's
why some devices will need to limit AVC codec profile level for screen
recorder only, as it doesn't make sense to do it globally, because it
will also affect camera and video editing.

Test: record the screen using the closest profile level to the display
      capabilities (resolution@frame rate). Check if video max. resolution and
      frame rate match the selected profile level constraints [1]. In case the
      profile level is the right one, the quality and fluidity of the video
      won't be in any way different from what you normally see on your display
      and will be playable anywhere, including the device that recorded it.

[1] https://en.wikipedia.org/wiki/Advanced_Video_Coding#Levels

Signed-off-by: iusmac <iusico.maxim@libero.it>
[@neobuddy89: Added default config to 4.2 and simplified!]
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Co-Authored-By: David Ramírez  <DavidYonatan9@hotmail.com>